refactored code: Basehandlertest and SetOptionTraitTest
